"Make site accessible for older browsers"
To make rem sizing simpler, I made the
base font size of the html element 6.25%.

But after some research, I found that the
convention of 62.5% is much better because
some older browsers will not understand
rem and will fallback on the percentage.

With text that small (6.25%) the site will
be unreadable on older browsers.